pyproject.toml
[project]
name = "snow"
description = "Static analysis tooling for Semgrep"
requires-python = "^3.6.2"
authors = [
    {"name" = "Erin Browning", "email" = "ebrowning@slack-corp.com"},
    {"name" = "Tim Faraci", "email" = "tfaraci@slack-corp.com"},
    {"name" = "Olivier Arteau", "email" = "oarteau@slack-corp.com"}
]

[tool.black]
line-length = 120
extend-exclude = 'languages/'

[tool.poetry]
name = "snow"
version = "0.1.0"
description = ""
authors = ["Product Security @ Slack"]

[tool.poetry.dependencies]
python = "^3.6.2"
PyYAML = "^6.0"
boto3 = "^1.21.29"
requests = "^2.27.1"
chardet = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = ">7.0"
black = ">0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
